+++ v2/src/SchoolManagement.Core/Models/UserModel.cs
@@ -1,0 +1,10 @@
+﻿namespace SchoolManagement.Core.Models;
+
+public record UserModel(
+    Guid Id,
+    string Username,
+    DateTime CreatedAt,
+    DateTime? UpdatedAt,
+    UserStatus Status,
+    List<string> Roles);
+
