+++ v2/src/SchoolManagement.Services.Identity/Entities/User.cs
@@ -1,4 +1,5 @@
 ﻿using SchoolManagement.Core.Entities;
+using SchoolManagement.Core.Models;
 
 namespace SchoolManagement.Services.Identity.Entities;
 
@@ -11,4 +12,12 @@ public class User : Entity, IAuditable
     public UserStatus Status { get; set; }
 
     public virtual ICollection<UserRole> Roles { get; set; } = null!;
+}
+public static class UserExtensions
+{
+    public static UserModel ToModel(this User user)
+    {
+        var roles = user.Roles?.Select(r => r.Role.Name).ToList() ?? new List<string>();
+        return new UserModel(user.Id, user.Username, user.CreatedAt, user.UpdatedAt, user.Status, roles);
+    }
 }
