+++ v2/src/SchoolManagement.Services.Students/Consumers/UserConsumer.cs
@@ -1,0 +1,20 @@
+﻿using MassTransit;
+using SchoolManagement.Core.Models;
+using SchoolManagement.Services.Students.Context;
+
+namespace SchoolManagement.Services.Students.Consumers;
+
+public class UserConsumer: IConsumer<UserModel>
+{
+    private StudentsDbContext _studentsDbContext;
+
+    public UserConsumer(StudentsDbContext studentsDbContext)
+    {
+        _studentsDbContext = studentsDbContext;
+    }
+
+    public async Task Consume(ConsumeContext<UserModel> context)
+    {
+        var userModel = context.Message;
+    }
+}
